dynamic max partsize
Before this it was hardcoded to 8mb. This was suboptimal, because:
1. upper bound often could be much stricter
2. in some files the samples are bigger (uncompressed 4k videos)

Currently the (new) stricter upper bound is only used in '-s' mode, or
if case 2. is detected. This should avoid beeing to resetrict when the
chances of false positives are relatively low.
However one can use '-mp f' to force the usage of the stricter upper bound.
It is also possible to supply a number, e.g. '-mp 9222' or '-mp 4mb'.